trying to make the plots easier
